feat: Add BaseSnackbar.tsx to reuse code
